broke out io connect into a module and removed code from server.js
